requirements.txt
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
dateparser==1.2.0
defusedxml==0.7.1
dnspython==2.7.0
filelock==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
markdownify==0.13.1
packaging==24.1
parsel==1.9.1
pluggy==1.5.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.10.1
pyOpenSSL==24.2.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
tldextract==5.1.2
Twisted==24.10.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.1.1
